pyproject.toml
[project]
name = "sikulipy"
version = "0.0.1"
description = "Python port of OculiX/SikuliX — visual automation with a Flet-based IDE."
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "Ihor Honcharov" }]
keywords = ["automation", "sikuli", "oculix", "opencv", "flet", "gui-testing"]
classifiers = [
    "Development Status :: 1 - Planning",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: User Interfaces",
]

# Core dependencies. Heavy / optional stacks are behind extras so the core
# install stays lean. See ROADMAP.md for the porting plan.
dependencies = [
    "flet>=0.25",
    "opencv-python>=4.10",
    "numpy>=2.0",
    "pillow>=10.0",
    "mss>=9.0",            # fast cross-platform screen capture
    "pyautogui>=0.9.54",   # mouse / keyboard fallback
    "pynput>=1.7",         # low-level input + global hotkeys
    "pydantic>=2.8",       # settings / options model
    "typer>=0.12",         # CLI entry point
    "rich>=13.7",          # console UX
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "paddleocr>=2.7",
]
android = [
    "pure-python-adb>=0.3.0.dev0",
]
vnc = [
    "vncdotool>=1.2",
]
runners = [
    "robotframework>=7.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
sikulipy = "sikulipy.cli:app"
sikulipy-ide = "sikulipy.ide.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sikulipy"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
